+++ lib/stack_master/commands/apply.rb
@@ -1,3 +1,5 @@
+require 'pathname'
+
 module StackMaster
   module Commands
     class Apply
@@ -208,8 +210,13 @@ module StackMaster
 
       def ensure_valid_parameters!
         if @proposed_stack.missing_parameters?
-          StackMaster.stderr.puts "Empty/blank parameters detected, ensure values exist for those parameters. Parameters will be read from the following locations:"
-          @stack_definition.parameter_files.each do |parameter_file|
+          StackMaster.stderr.puts <<~MESSAGE
+            Empty/blank parameters detected, ensure values exist for those parameters.
+            Parameters will be read from the following locations:
+          MESSAGE
+          base_dir = Pathname.new(@stack_definition.base_dir)
+          @stack_definition.parameter_file_globs.each do |glob|
+            parameter_file = Pathname.new(glob).relative_path_from(base_dir)
             StackMaster.stderr.puts " - #{parameter_file}"
           end
           halt!